requirements.txt
anyio==3.6.2
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
fastapi==0.95.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
python-dotenv==1.0.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
